Fixed #83
This problem could really be considered a third-party plugin support
fix. Some plugins, such as OneTimeUse, add glow to their items.
However, they use the traditional method, using NMS, to apply the glow
effect. This effect, because and NBT tag was edited, was not saved to
the config. The new method of applying the glow effect, using the
HIDE_ENCHANTS item flag, does save to the config, and does not require
NMS. To fix this bug, we simply had to check if the items are using the
old method of adding glow, apply the new method if so, and save the
item. This is a simple fix. There is one caveat: This does not seem to
be supported in 1.7. So if the server is a 1.7 server, then the glow
effect will not be saved.